requirements.txt
Django==1.11
Pillow==4.2
rarfile==3.0
